Start refactoring the browse scheme page, splitting into components. Also adjust the data shown in map tooltips
